some refactor

+++ app/Http/Controllers/Admin/ClearDataBaseController.php
@@ -3,31 +3,32 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Services\SettingsService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\File;
 
 class ClearDataBaseController extends Controller
 {
-    public function index() {
+    public function index()
+    {
         return view('Admin.Clear-DataBase.index');
     }
-    public function clearDatabase() {
-        try
-        {
+    public function clearDatabase()
+    {
+        try {
             Artisan::call('migrate:fresh');
-            Artisan::call('db:seed',['--class'=> 'MenuBuilderSeeder']);
-            Artisan::call('db:seed',['--class'=> 'PaymentGatewaySettingSeeder']);
-            Artisan::call('db:seed',['--class'=> 'SectionTitleSeeder']);
-            Artisan::call('db:seed',['--class'=> 'SettingSeeder']);
-            Artisan::call('db:seed',['--class'=> 'UserSeeder']);
+            Artisan::call('db:seed', ['--class' => 'MenuBuilderSeeder']);
+            Artisan::call('db:seed', ['--class' => 'PaymentGatewaySettingSeeder']);
+            Artisan::call('db:seed', ['--class' => 'SectionTitleSeeder']);
+            Artisan::call('db:seed', ['--class' => 'SettingSeeder']);
+            Artisan::call('db:seed', ['--class' => 'UserSeeder']);
 
             $this->deleteFiles();
-            return response(['status' => 'success','message' => 'Database is successfully cleared']);
-
-
-        }catch(\Exception $e)
-        {
+            $settingService = app(SettingsService::class);
+            $settingService->clearCachedSettings();
+            return response(['status' => 'success', 'message' => 'Database is successfully cleared']);
+        } catch (\Exception $e) {
             throw $e;
         }
     }
@@ -37,8 +38,9 @@ class ClearDataBaseController extends Controller
         $path = public_path('uploads');
         $allFiles = File::allFiles($path);
         foreach ($allFiles as $file) {
-            File::delete($file);
+            if (!str_contains($file->getPathname(), 'uploads\Default')) {
+                File::delete($file);
+            }
         }
     }
-
 }
